backend/requirements.txt
fastapi[standard]~=0.115.6

starlette~=0.41.3
python-jose~=3.3.0
passlib~=1.7.4
pydantic~=2.10.5
SQLAlchemy~=2.0.37